Exchanged int for size_t.

+++ include/okruz/bspline/operators/Position.h
@@ -44,7 +44,7 @@ class Position : public Operator {
   static std::array<T, n + 1> expandPower(const T &xm) {
     std::array<T, n + 1> retVal;
     T power_of_xm = static_cast<T>(1);
-    for (int i = 0; i < n + 1; i++) {
+    for (size_t i = 0; i < n + 1; i++) {
       retVal[n - i] = internal::binomialCoefficient<T>(n, i) * power_of_xm;
       power_of_xm *= xm;
     }
@@ -93,8 +93,8 @@ class Position : public Operator {
     std::array<T, OUTPUT_SIZE> retVal;
     retVal.fill(static_cast<T>(0));
 
-    for (int i = 0; i < input.size(); i++) {
-      for (int j = 0; j < expanded.size(); j++) {
+    for (size_t i = 0; i < input.size(); i++) {
+      for (size_t j = 0; j < expanded.size(); j++) {
         retVal[i + j] += expanded[j] * input[i];
       }
     }
